Implemented 'free ...'
The 'free' keyword is now supported, allowing for dynamically-allocated memory to be released back to the VM. This commit also comes along with various changes for code maintenance, such as moving  'Symbol' into its own file, modifying the types of exceptions thrown, reserving more keywords in the Lexer, and adding comment headers to source files.

+++ Compiler.h
@@ -49,7 +49,7 @@ class Compiler
 
 	size_t stack_offset;
 
-	unsigned int current_scope;	// tells us what scope level we are currently in
+	size_t current_scope;	// tells us what scope level we are currently in
 	std::string current_scope_name;
 
 	size_t strc_number;	// the next available number for a string constant
